requirements.txt
gspread>=5.8.0
google-api-python-client>=2.99.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
gspread-formatting>=1.1.7
requests>=2.31.0
pandas>=2.2.0